### 2020-03-01 #2130 @whedon
[PRE REVIEW]: Matching: A Python library for solving matching games
**Submitting author:** @daffidwilde (<a href="http://orcid.org/0000-0002-3788-7691">Henry Wilde</a>)
**Repository:** <a href="https://github.com/daffidwilde/matching" target ="_blank">https://github.com/daffidwilde/matching</a>
**Version:** v1.2.1
