### 2020-03-01 #2130 @labarba
👋  @VivianePons — The authors suggested you as handling editor for this submission. It's a package for solving matching games: games with two sets of players, where each player in one set has a ranked preference list of those in the other, and the objective is to find some mapping between the two sets such that no pair of players are unhappy.

Could you handle it?
