### 2020-03-23 #2169 @whedon
[REVIEW]: Matching: A Python library for solving matching games
**Submitting author:** @daffidwilde (<a href="http://orcid.org/0000-0002-3788-7691">Henry Wilde</a>)
**Repository:** <a href="https://github.com/daffidwilde/matching" target ="_blank">https://github.com/daffidwilde/matching</a>
**Version:** 1.3.0
