### 2020-03-26 #2169 @mbdemoraes
The authors propose a python library to solve four "matching games" problems. I could locate the repository, install and run the examples from both README and documentation. The paper is well-written and introduces the field clearly and objectively. Also, the "Discussion" section in the documentation allows the reader to understand the subject in greater depth. Good job!

Some minor aspects that need revision on the documentation:

- There is a "How to" page that is empty;
- Under "Indice and Tables" on the home page, there is a "Search page" that is also empty;

Also, I have a few suggestions that I believe will improve this work:

- Remove the "code figure" from the paper. Although JOSS is a software-oriented journal, the paper is the place where you introduce motivations, theoretical and scientific aspects of your work. Software functionality should be addressed only on the documentation;

- Include on the documentation that for people with two python installations (versions 2 and 3+), they may have to use the command "python3" to install the package and run the examples, since in that case, the "python" command is the default for python2. It would be a command like this: "python3 -m pip install matching" and to run an example "python3 example.py". This is a common problem when you do not have a scientific python distribution (such as anaconda) installed. It may be obvious for experienced users but others may experience some difficulties;

- In the paper, the first paragraph of the "Statement of Need" section, the authors write that "Matching games have applications in a number of fields including social and market economics, wireless communication and education infrastructure". Although a practical application in education infrastructure is easily understandable given the examples itself (for instance, the student-allocation problem), applications in market economics and wireless communication are not so easily perceptible. If the authors can provide some specific applications under these fields, more researchers can understand the applicability of the library.

- On the "Get in contact!" section of README, the authors use "PRs" as an acronym for "Pull requests". I know it is a well-known abbreviation for GitHub users, but again, some readers may not understand. I suggest using the full text "Pull requests" to improve readability.

Thank you for the opportunity to review this work! Let me know when the revision has been made.
