### 2020-03-29 #2169 @igarizio
👋 Hi @daffidwilde. Thank you for submitting your paper!

The paper is well written and the code is clear, easy to use and properly documented (good job!). So far, I have only found some small issues:
- As @mbdemoraes  mentioned, the “How to” page on the documentation is empty.
- In the setup.py file, I think you may be missing the hypothesis package (in “tests_require”). (The package is correctly listed everywhere else.) After installing it all tests pass. I should also mention that the coverage is perfect!
- I get a ValueError when running a StableMarriage game twice. I opened an issue with more details [daffidwilde/matching#80].

Opinion:
- While the “code figure” on the paper may look a little bulky, I do not think you need to remove it, as it highlights how easy it is to use your package (maybe you can remove the “>“ (?)).
@mbdemoraes It is true that JOSS says not to include “software documentation such as API functionality” ([link](https://joss.readthedocs.io/en/latest/submitting.html)), but I think in this case, it is trying to show a strength of the package, not documenting it. What do you think? 👀
- You mention that the Game class is just a “base class and is not intended for use other than inheritance”. Maybe you can use a metaclass to highlight that (so users get a nice warning on their IDE when subclassing it). This is just a suggestion, your code in completely fine without it.
