### 2020-03-29 #2169 @mbdemoraes
@igarizio Thank you! I suggested that because I thought "code in the paper" was a restriction. But if it's not, you are certainly right, the image sure adds to the content. :smile:
