### 2020-03-29 #2169 @igarizio
@mbdemoraes Perfect! As a reference, JOSS even has an example on how to add python code [here](https://joss.readthedocs.io/en/latest/submitting.html), and also there are other accepted papers with code in them (for example [this one](https://joss.theoj.org/papers/10.21105/joss.01882)).
