### 2020-03-31 588faed
Remove duplicate "Documentation" sect from README
