### 2020-03-31 2a3d764
Implement player copying.
Player instances are copied at the instantiation of a Game subclass.
This is done using `copy.deepcopy`. That way, multiple Game instances
can be created from one set of player lists.
