### 2020-03-31 #2169 @daffidwilde
Hello everyone,

Thank you so much for getting back with your reviews so swiftly. I've opened a [number of PRs](https://github.com/daffidwilde/matching/pulls) on the matching repository that address most of the points you've raised and I'd appreciate your feedback.

I will be adding to these tomorrow but I think things are looking pretty good thus far.

Thanks :+1:
