### 2020-03-31 #2169 @daffidwilde
One thing I haven't yet changed that I'd like to discuss more is this point raised by @mbdemoraes:

> Include on the documentation that for people with two python installations (versions 2 and 3+), they may have to use the command "python3" to install the package and run the examples, since in that case, the "python" command is the default for python2. It would be a command like this: "python3 -m pip install matching" and to run an example "python3 example.py". This is a common problem when you do not have a scientific python distribution (such as anaconda) installed. It may be obvious for experienced users but others may experience some difficulties;

When writing the documenation, we decided that since Python 2 is not supported in Matching (by the use of f-strings, for instance) that this was not necessary -- although I see why that could become an issue as you've highlighted. Having said that, since Python 2 has been officially sunsetted as of January 2020, I would prefer to avoid distinguishing between "Python" and "Python 3".

Perhaps adding this requirement to the `setup.py` and making it clearer (via a warning directive) in the docs would be sufficient? We can discuss this more on [this PR](https://github.com/daffidwilde/matching/pull/85) if you'd like so as not to clutter this issue.
