### 2020-03-31 #2169 @mbdemoraes
@daffidwilde I understand! Adding this requirement to the setup.py and/or making it clearer in the documentation would sure be sufficient. :thumbsup:
