### 2020-04-02 fefe851
Add `ValueError` to exceptions in SR(2) test.
Since the first phase is not being completed, the preferences have not
been reduced and the all-or-nothing cycles can contain repeats of pairs.
This means players attempt to forget someone they've already forgotten.
