### 2020-04-02 50c14ec
Remove the try-except from `Player.forget()`.
No longer necessary as the tests catch this edge case (that would not
ever actually happen in SR).
