### 2020-04-02 daffidwilde/matching#90 @daffidwilde
Should `matching.Game` be a metaclass?
On [the review](https://github.com/openjournals/joss-reviews/issues/2169#issuecomment-605560287), @igarizio raised the following point about `matching.Game`:

> On game.py, you mention that the Game class is just a “base class and is not intended for use other than inheritance”. Maybe you can use a metaclass to highlight that (so users get a nice warning on their IDE when subclassing it). This is just a suggestion, your code in completely fine without it.

I was wondering how this might be done. My initial attempt has been to let `matching.Game` inherit from `type` but that would mean a large refactor of the inheritance for the rest of the games since a metaclass has its own requisites beyond a class.

Is there a simpler way that I'm totally missing? I agree it'd be nice to give users some feedback when using the library to implement new games but don't want to cloud the architecture too much.
