+++ INSTALLATION.rst
@@ -1,13 +1,13 @@
-Matching is written in Python 3, and relies only on `NumPy
-<http://www.numpy.org/>`_ for general use.
+Matching requires a Python version of 3.5 or above and relies only on `NumPy
+<http://www.numpy.org/>`_ for installation and general use.
 
 The library is most easily installed using :code:`pip`::
 
     $ python -m pip install matching
 
 However, if you would like to install it from source then go ahead and clone the
-GitHub repo::
+GitHub repository before installing locally::
 
     $ git clone https://github.com/daffidwilde/matching.git
     $ cd matching
-    $ python setup.py install
+    $ python -m pip install .
