### 2020-04-02 1bfb9b9
Move install intro to a note directive.
