### 2020-04-02 5605770
Remove fluff around edu infrastructure.
