### 2020-04-03 daffidwilde/matching#90 @igarizio
@daffidwilde Thank you for taking the time to consider the suggestion!
My go-to solution for metaclasses is through [abc](https://docs.python.org/3/library/abc.html) (part of the standard library). Here is how ``Game`` would look like using ``abc``:

```python
import abc

class BaseGame(metaclass=abc.ABCMeta):  # Renamed just to make it more clear.

    def __init__(self):
        self._matching = None
        self.blocking_pairs = None

    @property
    def matching(self):
        """ Property method to stop direct write access. """
        return self._matching

    @matching.getter
    def matching(self):
        """ Property method to stop direct write access. """
        return self._matching

    @abc.abstractmethod
    def solve(self):
        """ Placeholder for solving the given matching game. """
        pass

    @abc.abstractmethod
    def check_stability(self):
        """ Placeholder for checking the stability of the current matching. """
        pass

    @abc.abstractmethod
    def check_validity(self):
        """ Placeholder for checking the validity of the current matching. """
        pass
```
Now, if I try to instantiate ``BaseGame`` I get ``TypeError: Can't instantiate abstract class BaseGame with abstract methods...``. Also, if I subclass it without implementing the required methods, my IDE (Pycharm) says "Class Dummy must implement all abstract methods". Finally, all tests pass, except for the ones on test_game.py, which is expected since it's not possible to instantiate ``Game`` anymore. (If I may suggest a way rewrite those tests, I would say that creating a dummy subclass and test the parameters there, would be a good idea.)
