### 2020-04-03 119ee89
Change inheritance in each game.
