+++ src/matching/games/stable_marriage.py
@@ -1,11 +1,11 @@
 """ The SM solver and algorithm. """
 
-from matching import Game, Matching, Player
+from matching import BaseGame, Matching, Player
 
 from .util import delete_pair, match_pair
 
 
-class StableMarriage(Game):
+class StableMarriage(BaseGame):
     """ A class for solving instances of the stable marriage problem (SM).
 
     Parameters
