### 2020-04-03 7da3dc1
Remove `matching` property from `Game`.
