### 2020-04-03 ad3f412
Remove `matching` property (#95)
* Remove `matching` property from `Game`.

* Make `matching` a normal attribute in each game.
