+++ src/matching/games/stable_roommates.py
@@ -40,7 +40,7 @@ class StableRoommates(Game):
         """ Solve the instance of SR using Irving's algorithm. Return the
         matching. """
 
-        self._matching = Matching(stable_roommates(self.players))
+        self.matching = Matching(stable_roommates(self.players))
         return self.matching
 
     def check_stability(self):
