### 2020-04-03 fabb94f
Fix typo in game docstring and matching test. (#96)
