### 2020-04-03 #2169 @daffidwilde
👋 Hello all!

I've now completed making the changes requested here in Matching. The changes have been split amongst [seven pull requests](https://github.com/daffidwilde/matching/pulls).

It'd be great if you could have a look at these changes whenever you get the chance.

Thanks again for all the suggestions and discussion 😄
