### 2020-04-06 330f03b
Add hypothesis to setup.py (#87)
* Add test for inconsistent coverage. (#86)

Occasionally, coverage falls below 100% because of an edge case in HR
where no resident is removed from consideration for forgetting all of
their preferences. This simple example test rectifies that.

* Add hypothesis to `tests_require` in `setup.py`
