### 2020-04-06 ab96b62
Make Game a metaclass (#94)
* Reimplement `test_game.py` for `abc.ABCMeta`.

* Change `Game` to `BaseGame` as `ABCMeta` subclass.

* Change inheritance in each game.
