### 2020-04-08 2215dbe
Fix documentation for JOSS review (#85)
* Add sphinxcontrib-bibtex to docs/env file.

* Remove "Search Page" from "Indices and Tables"

* Add notebook pointer prologue.

* Clarify installation instructions for pyversion.

* Move install intro to a note directive.
