### 2020-04-08 f3fe69d
Update README (and supporting files) (#84)
* Add CONTRIBUTING file to the top level.

* Remove duplicate "Documentation" sect from README

* Change "PR" to "pull request" in README
