+++ docs/reference/contributing.rst
@@ -1,28 +1,4 @@
 Contributing to the library
 ===========================
 
-If you are interested in contributing to the Matching library then first of all
-we would like to offer our thanks and say welcome!
-
-Whether you're looking to provide a fix for a current `issue
-<https://github.com/daffidwilde/matching/issues>`_, report a bug or implement a
-new feature to the library, your contributions are always welcome. Pull requests
-(PRs) are a good place to discuss contributions so please do not feel you must
-present a perfect product from the off.
-
-To make a contribution via a PR, follow these steps:
-
-1. Go to the `GitHub repo <https://github.com/daffidwilde/matching>`_,
-   make a fork, and clone the repo locally::
-
-       $ git clone git@github.com/<your-username>/matching.git
-
-2. Assuming you have all the dependencies installed, the next step is to ensure
-   that all the tests pass (this should not take very long at all)::
-
-       $ cd matching
-       $ pytest tests
-
-3. Make your changes and write tests to go with them -- ensuring they pass, too.
-
-4. Push to your fork and open a pull request.
+.. include:: ../../CONTRIBUTING.rst
