### 2020-04-08 c34f90a
Add python_requires parameter to setup (#92)
