### 2020-04-14 #2169 @igarizio
@daffidwilde I may be wrong, but I think some of the changes need to be merged into the master branch. I think I can still generate the new pdf from your joss-review branch, but I'm not sure if whedon can check the references there.
