+++ paper.md
@@ -84,8 +84,10 @@ require the use of software to be solved in reasonable time.
 
 # Statement of Need
 
-Matching games have applications in a number of fields including social and
-market economics [@agar2017; @okun1995], wireless communication [@baya2016], and
+Matching games have applications in many fields where relationships between
+rational agents must be managed. Some example applications include: being able
+to inform on healthcare finance policy [@agar2017]; helping to reduce the
+complexity of automated wireless communication networks [@baya2016]; and
 education infrastructure [@chia2019]. Thus, having access to software
 implementations of algorithms that are able to solve such games is essential.
 
