+++ paper.bib
@@ -114,18 +114,6 @@
     note = {[Online; accessed 2020-02-24]},
 }
 
-@article{okun1995,
-    title = {Social Norms and Random Matching Games},
-    journal = {Games and Economic Behavior},
-    volume = {9},
-    number = {1},
-    pages = {79 - 109},
-    year = {1995},
-    issn = {0899-8256},
-    doi = {10.1006/game.1995.1006},
-    author = {Okuno-Fujiwara, Masahiro and Postlewaite, Andrew},
-}
-
 @article{roth1984,
     author = {Roth, Alvin},
     title = {The Evolution of the Labor Market for Medical Interns and
