+++ docs/tutorials/project_allocation/data.py
@@ -188,9 +188,10 @@ def main():
     print("Seed set:", SEED)
 
     supervisor_to_projects = create_supervisor_to_projects_map()
-    supervisor_to_capacity, project_to_capacity = create_player_to_capacity_maps(
-        supervisor_to_projects
-    )
+    (
+        supervisor_to_capacity,
+        project_to_capacity,
+    ) = create_player_to_capacity_maps(supervisor_to_projects)
     print("Supervisor and project dictionaries created...")
 
     all_projects = list(get_all_projects(supervisor_to_projects))
