### 2020-04-14 765e82e
Update citation info in paper bib
