+++ paper.bib
@@ -98,11 +98,11 @@
 }
 
 @misc{matching,
-   author        = {{ {The Matching library developers} }},
-   title         = {Matching: v1.2.1},
+   author        = {{The Matching library developers}},
+   title         = {Matching: v1.3},
    year          = {2020},
-   doi           = {10.5281/zenodo.3691470},
-   url           = {http://doi.org/10.5281/zenodo.3691470}
+   doi           = {10.5281/zenodo.3751325},
+   url           = {http://doi.org/10.5281/zenodo.3751325}
 }
 
 @misc{numpy,
