### 2020-04-14 #2169 @VivianePons
Thanks @arfon I can handle it! I see that both reviewers are happy so we can move forward.

@daffidwilde can you give a final check and then create an archive (on Zenodo, figshare, or other) and send me the ref. I will have time to make a final pass on Friday and proceed with acceptance. Congrats! And thanks @mbdemoraes and @igarizio for the reviews
