### 2020-04-17 #2169 @drvinceknight
DOI is `10.5281/zenodo.3755304`
Version is `v1.3.0`
